Move worker from example into rusie_queue src
