+++ examples/hello.rs
@@ -1,24 +1,15 @@
-use futures::{stream, StreamExt};
 use rusie_queue::error::Error;
-use rusie_queue::postgres::*;
-use serde::de::DeserializeOwned;
+use rusie_queue::postgres::{PostgresJob, PostgresQueue};
+use rusie_queue::worker;
 use serde::{Deserialize, Serialize};
 use sqlx::postgres::PgPoolOptions;
 use sqlx::{Pool, Postgres};
-use std::fmt::{Debug, Display};
-use std::future::Future;
 use std::sync::Arc;
 use std::time::Duration;
 
-const CONCURRENCY: usize = 50;
-
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Serialize, Deserialize)]
 pub enum Message {
-    SendSignInEmail {
-        email: String,
-        name: String,
-        code: String,
-    },
+    SayHello { name: String },
 }
 
 #[tokio::main]
@@ -31,76 +22,24 @@ async fn main() -> Result<(), anyhow::Error> {
 
     let queue = Arc::new(PostgresQueue::new(db.clone()).max_attempts(3));
 
-    // run worker
-    let worker_queue = queue.clone(); // queue is an Arc pointer, so we only copy the reference
-    tokio::spawn(async move { run_worker(worker_queue, handle_job).await });
+    let worker_queue = queue.clone();
+    tokio::spawn(async move { worker::run(worker_queue, handle_job).await });
 
-    // queue job
-    let job = Message::SendSignInEmail {
-        email: "your@email.com".to_string(),
+    let job = Message::SayHello {
         name: "Rusie Q".to_string(),
-        code: "000-000".to_string(),
     };
-    let _ = queue.push(job, None).await; // TODO: handle error
+
+    queue.push(job, None).await?;
 
     tokio::time::sleep(Duration::from_secs(2)).await;
 
     Ok(())
 }
 
-// TODO: move into rusie_queue
-async fn run_worker<M, E, F, Fut>(queue: Arc<PostgresQueue>, handler: F)
-where
-    M: Serialize + DeserializeOwned + Unpin + Send + 'static,
-    E: Display,
-    F: Fn(PostgresJob<M>) -> Fut + Send + Sync + 'static,
-    Fut: Future<Output = Result<(), E>>,
-{
-    loop {
-        let jobs = match queue.pull::<M>(CONCURRENCY as u32).await {
-            Ok(jobs) => jobs,
-            Err(err) => {
-                println!("run_worker: pulling jobs: {}", err);
-                tokio::time::sleep(Duration::from_millis(500)).await;
-                Vec::new()
-            }
-        };
-
-        let number_of_jobs = jobs.len();
-        if number_of_jobs > 0 {
-            println!("Fetched {} jobs", number_of_jobs);
-        }
-
-        stream::iter(jobs)
-            .for_each_concurrent(CONCURRENCY, |job| async {
-                let job_id = job.id;
-
-                let res = match handler(job).await {
-                    Ok(_) => queue.delete_job(job_id).await,
-                    Err(err) => {
-                        println!("run_worker: handling job({}): {}", job_id, &err);
-                        queue.fail_job(job_id).await
-                    }
-                };
-
-                match res {
-                    Ok(_) => {}
-                    Err(err) => {
-                        println!("run_worker: deleting / failing job: {}", &err);
-                    }
-                }
-            })
-            .await;
-
-        // sleep not to overload our database
-        tokio::time::sleep(Duration::from_millis(125)).await;
-    }
-}
-
 async fn handle_job(job: PostgresJob<Message>) -> Result<(), anyhow::Error> {
     match job.message.0 {
-        message @ Message::SendSignInEmail { .. } => {
-            println!("Sending sign in email: {:?}", &message);
+        Message::SayHello { name } => {
+            println!("Hello, {}", name);
         }
     };
