+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod error;
 pub mod postgres;
+pub mod worker;
